// Portal SAE/frontend/src/app/interfaces/contact/contact.dto.ts
export interface ContactDTO {
  codigo: number;
  numeroIdentificacion: string;
  nombres: string;
  apellidos: string;
  pais: string;
  departamento: string;
  ciudad: string;
  direccion: string;
  codigoPais: string;
  movil: string;
  email: string;
  fechaNacimiento: string;
  rutaImagen: string;
  tipoIdentificacion: string;
  cuenta: string;
}