// Portal SAE/frontend/src/app/interfaces/contact/createContact.dto.ts
export interface CreateContactDTO {
  numeroIdentificacion: string;
  nombres: string;
  apellidos: string;
  pais: string;
  departamento: string;
  ciudad: string;
  direccion: string;
  codigoPais: string;
  movil: string;
  email: string;
  fechaNacimiento: string;
  rutaImagen: string;
  tipoIdentificacion: string;
  cuenta: string;
}