// Portal SAE/frontend/src/app/interfaces/contact/updateContact.dto.ts
export interface UpdateContactDTO {
  numeroIdentificacion?: string;
  nombres?: string;
  apellidos?: string;
  pais?: string;
  departamento?: string;
  ciudad?: string;
  direccion?: string;
  codigoPais?: string;
  movil?: string;
  email?: string;
  fechaNacimiento?: string;
  rutaImagen?: string;
  tipoIdentificacion?: string;
  cuenta?: string;
}